home_gate/bot/src/requirements.txt
python-telegram-bot
paho-mqtt
Flask
